Add files via upload
This is the code I used to gather most of the preliminary results for my research symposium. I am still working on this research so I plan on updating this code a lot with probably later ones being labeled with a date of which they were released. I also provide the poster I made and presented at the symposium. The pickled data will also be provided with the code.